+++ udata_ods/__init__.py
@@ -3,5 +3,5 @@ OpenDataSoft integration for udata
 '''
 from __future__ import unicode_literals
 
-__version__ = '3.0.1'
+__version__ = '3.0.2.dev'
 __description__ = 'OpenDataSoft integration for udata'