+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- :warning: Add migration from ODS to DCAT [#247](https://github.com/opendatateam/udata-ods/pull/247)
+    - All ODS harvest sources are converted to DCAT
+    - Datasets are updated with new remote_id and new resources urls
 
 ## 3.0.1 (2023-02-06)